Compain api get and send functions in one js
