+++ src/index.js
@@ -1,6 +1,5 @@
 import './style.css';
-import { populate } from './populate.js';
-import { broadcast } from './broadcast.js';
+import { broadcast, populate } from './broadcast.js';
 
 // Update scores when refresh button is pressed
